jimt/log-js-errors - properly updates properties
