+++ apps/src/lab2/views/Lab2Wrapper.tsx
@@ -16,6 +16,8 @@ import FontAwesome from '@cdo/apps/templates/FontAwesome';
 import Lab2MetricsReporter from '../Lab2MetricsReporter';
 const i18n = require('@cdo/locale');
 
+import {ErrorFallbackPage, ErrorUI} from './ErrorFallbackPage';
+
 export interface Lab2WrapperProps {
   children: React.ReactNode;
 }
@@ -73,30 +75,5 @@ const Lab2Wrapper: React.FunctionComponent<Lab2WrapperProps> = ({children}) => {
     </ErrorBoundary>
   );
 };
-
-export interface ErrorUIProps {
-  message?: string;
-}
-
-export const ErrorUI: React.FunctionComponent<ErrorUIProps> = ({message}) => (
-  <div id="page-error-container" className={moduleStyles.pageErrorContainer}>
-    <div id="page-error" className={moduleStyles.pageError}>
-      <img
-        className={moduleStyles.pageErrorImage}
-        src="/shared/images/sad-bee-avatar.png"
-      />
-      <div>{i18n.loadingError()}</div>
-      {message && (
-        <div className={moduleStyles.pageErrorMessage}>({message})</div>
-      )}
-    </div>
-  </div>
-);
-
-export const ErrorFallbackPage = () => (
-  <div id="lab-container" className={moduleStyles.labContainer}>
-    <ErrorUI />
-  </div>
-);
 
 export default Lab2Wrapper;
